requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
cryptography==39.0.2
Deprecated==1.2.13
Django==4.1.7
django-oauth-toolkit==2.2.0
djangorestframework==3.14.0
idna==3.4
jwcrypto==1.4.2
mo==0.3.0
mysql-connector-python==8.0.32
mysqlclient==2.1.1
oauthlib==3.2.2
pycparser==2.21
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
sqlparse==0.4.3
toml==0.10.2
tzdata==2022.7
urllib3==1.26.14
wrapt==1.15.0
